+++ src/atm_phy_schemes/mo_2mom_mcrph_driver.f90
@@ -242,7 +242,6 @@ CONTAINS
     ! as they are used in the various options of the scheme
     TYPE(particle), target   :: cloud_hyd, rain_hyd
     TYPE(particle), target   :: ice_frz, snow_frz, graupel_frz, hail_frz
-    TYPE(particle), target   :: graupel_lwf, hail_lwf
 
     ! Pointers to the derived types that are actually needed
     TYPE(particle), pointer  :: cloud, rain, ice, snow, graupel, hail
@@ -262,7 +261,7 @@ CONTAINS
     ENDIF
 #endif
 
-    IF (msg_level>5) CALL message (TRIM(routine), "called two_moment_mcrph")
+    IF (msg_level>10) CALL message (TRIM(routine), "called two_moment_mcrph")
 
     IF (PRESENT(ithermo_water)) THEN
        lconstant_lh = (ithermo_water == 0)
@@ -280,14 +279,8 @@ CONTAINS
     rain  => rain_hyd
     ice   => ice_frz
     snow  => snow_frz
-
-    IF (lprogmelt) THEN
-       graupel => graupel_lwf   ! with prognostic melting
-       hail => hail_lwf         ! of graupel and hail
-    ELSE
-       graupel => graupel_frz   ! simple melting
-       hail => hail_frz
-    END IF
+    graupel => graupel_frz
+    hail => hail_frz
 
     ! start/end indices
     IF (PRESENT(is)) THEN
@@ -435,7 +428,7 @@ CONTAINS
        IF (timers_level > 10) CALL timer_start(timer_phys_2mom_proc)
        ! .. this subroutine calculates all the microphysical sources and sinks
        CALL clouds_twomoment(ik_slice, dt, lprogin, &
-            atmo, cloud, rain, ice, snow, graupel, hail, ninact, nccn, ninpot)
+            atmo, cloud, rain, ice, snow, graupel, hail, ninact, nccn, ninpot, msg_level)
 
        IF (timers_level > 10) CALL timer_stop(timer_phys_2mom_proc)
 
@@ -670,7 +663,7 @@ CONTAINS
 
         ! .. this subroutine calculates all the microphysical sources and sinks
         CALL clouds_twomoment(ik_slice, dt, lprogin, atmo, cloud, rain, &
-             ice, snow, graupel, hail, ninact, nccn, ninpot)
+             ice, snow, graupel, hail, ninact, nccn, ninpot, msg_level)
 
         DO kk=kts,kte
           DO ii = its, ite
@@ -815,10 +808,8 @@ CONTAINS
         CALL sedi_vel_sphere(ice,ice_coeffs,qi(:,k),xi_now,rhocorr(:,k),vi_sedn_now,vi_sedq_now,its,ite)
         CALL sedi_vel_sphere(snow,snow_coeffs,qs(:,k),xs_now,rhocorr(:,k),vs_sedn_now,vs_sedq_now,its,ite)
         IF (lprogmelt) THEN
-          CALL sedi_vel_lwf(graupel_lwf,  &
-               & qg(:,k),qgl(:,k),xg_now,rhocorr(:,k),vg_sedn_now,vg_sedq_now,vg_sedl_now,its,ite)
-          CALL sedi_vel_lwf(hail_lwf,        &
-               & qh(:,k),qhl(:,k),xh_now,rhocorr(:,k),vh_sedn_now,vh_sedq_now,vh_sedl_now,its,ite)
+          CALL sedi_vel_lwf(graupel,qg(:,k),qgl(:,k),xg_now,rhocorr(:,k),vg_sedn_now,vg_sedq_now,vg_sedl_now,its,ite)
+          CALL sedi_vel_lwf(hail,qh(:,k),qhl(:,k),xh_now,rhocorr(:,k),vh_sedn_now,vh_sedq_now,vh_sedl_now,its,ite)
         ELSE
           CALL sedi_vel_sphere(graupel,graupel_coeffs,qg(:,k),xg_now,rhocorr(:,k),vg_sedn_now,vg_sedq_now,its,ite)
           CALL sedi_vel_sphere(hail,hail_coeffs,qh(:,k),xh_now,rhocorr(:,k),vh_sedn_now,vh_sedq_now,its,ite)
@@ -863,7 +854,7 @@ CONTAINS
           ik_slice(4) = k
           CALL clouds_twomoment(ik_slice, dt, lprogin, &
                atmo, cloud, rain, ice, snow, graupel, hail, &
-               ninact, nccn, ninpot)
+               ninact, nccn, ninpot, msg_level)
 
           ! .. latent heat term for temperature equation
           !$ACC PARALLEL ASYNC(1) DEFAULT(PRESENT)
@@ -1098,7 +1089,7 @@ CONTAINS
       ENDIF
        IF (lprogmelt) THEN
          DO ii=1,ntsedi_graupel
-           call sedi_icon_sphere_lwf(graupel_lwf,graupel_coeffs,qg,qng,qgl,&
+           call sedi_icon_sphere_lwf(graupel,graupel_coeffs,qg,qng,qgl,&
                 &                    prec_g,prec3D_tmp,rhocorr,rdz,dt/ntsedi_graupel,its,ite,kts,kte,cmax)
          END DO
        ELSE
@@ -1134,7 +1125,7 @@ CONTAINS
       ENDIF
       IF (lprogmelt) THEN
         DO ii=1,ntsedi_hail
-          call sedi_icon_sphere_lwf(hail_lwf,hail_coeffs,qh,qnh,qhl,&
+          call sedi_icon_sphere_lwf(hail,hail_coeffs,qh,qnh,qhl,&
                &                    prec_h,prec3D_tmp,rhocorr,rdz,dt/ntsedi_hail,its,ite,kts,kte,cmax)
         END DO
       ELSE
@@ -1306,7 +1297,6 @@ CONTAINS
          & N_in0,z0_nin,z1e_nin
 
     TYPE(particle) :: cloud, rain, ice, snow, graupel, hail
-    TYPE(particle) :: graupel_lwf, hail_lwf
 
     TYPE(t_cfg_2mom), OPTIONAL, INTENT(in) :: cfg_2mom
 
@@ -1338,9 +1328,9 @@ CONTAINS
 
     ! .. set the particle types, and calculate some coefficients
     IF (igscp == 7) THEN
-       CALL init_2mom_scheme_once(cloud,rain,ice,snow,graupel_lwf,hail_lwf,cloud_type)
+       CALL init_2mom_scheme_once(cloud,rain,ice,snow,graupel,hail,cloud_type,msg_level)
     ELSE
-       CALL init_2mom_scheme_once(cloud,rain,ice,snow,graupel,hail,cloud_type)
+       CALL init_2mom_scheme_once(cloud,rain,ice,snow,graupel,hail,cloud_type,msg_level)
     END IF
 
     IF (timers_level > 10) CALL timer_start(timer_phys_2mom_dmin_init)
@@ -1377,10 +1367,12 @@ CONTAINS
       N_cn0    = ccn_coeffs%Ncn0
     END IF
 
-    WRITE(message_text,'(A)') "  CN properties:" ; CALL message(TRIM(routine),TRIM(message_text))
-    WRITE(message_text,'(A,D10.3)') "    Ncn0 = ",ccn_coeffs%Ncn0 ; CALL message(TRIM(routine),TRIM(message_text))
-    WRITE(message_text,'(A,D10.3)') "    z0   = ",ccn_coeffs%z0  ; CALL message(TRIM(routine),TRIM(message_text))
-    WRITE(message_text,'(A,D10.3)') "    z1e  = ",ccn_coeffs%z1e ; CALL message(TRIM(routine),TRIM(message_text))
+    IF (msg_level > 7) THEN
+      WRITE(message_text,'(A)') "  CN properties:" ; CALL message(TRIM(routine),TRIM(message_text))
+      WRITE(message_text,'(A,D10.3)') "    Ncn0 = ",ccn_coeffs%Ncn0 ; CALL message(TRIM(routine),TRIM(message_text))
+      WRITE(message_text,'(A,D10.3)') "    z0   = ",ccn_coeffs%z0  ; CALL message(TRIM(routine),TRIM(message_text))
+      WRITE(message_text,'(A,D10.3)') "    z1e  = ",ccn_coeffs%z1e ; CALL message(TRIM(routine),TRIM(message_text))
+    END IF
 
     IF (PRESENT(N_in0)) THEN
 
@@ -1415,7 +1407,6 @@ CONTAINS
     ! These are the fundamental hydrometeor particle variables for the two-moment scheme
     TYPE(particle) , TARGET   :: cloud_hyd, rain_hyd
     TYPE(particle) , TARGET   :: ice_frz, snow_frz, graupel_frz, hail_frz
-    TYPE(particle) , TARGET   :: graupel_lwf, hail_lwf
 
     ! Pointers to the derived types that are actually needed
     TYPE(particle), POINTER  :: cloud, rain, ice, snow, graupel, hail
@@ -1440,13 +1431,9 @@ CONTAINS
     rain  => rain_hyd
     ice   => ice_frz
     snow  => snow_frz
-    IF (reff_calc%microph_param == 7 ) THEN  ! Vivek Param. Frozen+Liquid
-       graupel => graupel_lwf                ! gscp=7
-       hail    => hail_lwf
-    ELSE
-       graupel => graupel_frz                ! gscp=4,5,6
-       hail    => hail_frz
-    END IF
+    graupel => graupel_frz
+    hail  => hail_frz
+
     ! .. set the particle types, but no calculations
     CALL init_2mom_scheme(cloud,rain,ice,snow,graupel,hail)